+: insertion sort refactor
