+++ src/hooks/useIndexer/index.ts
@@ -1,2 +1,3 @@
 export { default } from "./lib/useIndexer";
 export * from "./lib/types";
+export * from "./lib/environmentStore";
