+++ src/page/view/InsertionSort/index.ts
@@ -1,3 +1,1 @@
 export {default} from "./ui";
-
-// LEFT OFF: 이제 로직을 어떻게 분리해낼지 고민해야대...
