+++ src/page/view/InsertionSort/model/store.ts
